refactor(core): unify KMP architecture with symmetric androidMain platform implementations
