+++ app/src/main/java/com/vinnovateit/latch/features/onboarding/pages/AccountPage.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.vinnovateit.latch.R
-import com.vinnovateit.latch.data.StoredCredentials
+import com.vinnovateit.latch.core.platform.android.StoredCredentials
 import com.vinnovateit.latch.features.onboarding.components.SlideContent
 import com.vinnovateit.latch.ui.theme.SatoshiFontFamily
