+++ core/src/androidMain/kotlin/com/vinnovateit/latch/core/platform/android/AndroidDatabaseFactory.kt
@@ -1,4 +1,4 @@
-package com.vinnovateit.latch.platform
+package com.vinnovateit.latch.core.platform.android
 
 import android.content.Context
 import androidx.room.Room
