+++ core/src/androidMain/kotlin/com/vinnovateit/latch/core/platform/android/AndroidKeyValueStore.kt
@@ -1,4 +1,4 @@
-package com.vinnovateit.latch.platform
+package com.vinnovateit.latch.core.platform.android
 
 import android.content.Context
 import androidx.core.content.edit
